pyproject.toml
[project]
name = "resver"
dynamic = ["version"]
description = "Model Group Manager — CLI lock-file registry for ML model versions"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer[all]>=0.12",
    "ruamel-yaml>=0.18",
    "semver>=3.0",
]

[project.scripts]
resver = "resver.cli:app"

[dependency-groups]
dev = [
    "mkdocs-material>=9.7.5",
    "pytest>=8.0",
    "pytest-mock>=3.14",
]

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "src/resver/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["src/resver"]

[tool.uv]
constraint-dependencies = [
  "click<=8.2.1"
]